Moved sound interfaces to sound/interfaces/ and renamed GOSoundOrganInterface* to GOSoundSamplerPlayer*
- Moved GOSoundCallbackConnector.h/.cpp, GOSoundCloseListener.h to sound/interfaces/
- Renamed GOSoundOrganInterface -> GOSoundSamplerPlayerInterface
- Renamed GOSoundOrganInterfaceProxy -> GOSoundSamplerPlayerProxy
- Updated all includes and class references accordingly
