+++ src/grandorgue/GOOrganController.cpp
@@ -846,7 +846,7 @@ void GOOrganController::StartOrgan(
 
   m_MidiSamplesetMatch.clear();
   GOOrganModel::SetMidi(&midi, m_MidiRecorder);
-  GOOrganModel::GOSoundOrganInterfaceProxy::Connect(
+  GOOrganModel::GOSoundSamplerPlayerProxy::Connect(
     &m_SoundEngine.GetSamplerPlayer());
   GOEventDistributor::PreparePlayback();
 
@@ -873,7 +873,7 @@ void GOOrganController::StopOrgan(GOSoundSystem &soundSystem) {
   m_AudioRecorder->SetAudioRecorder(NULL);
   if (p_OnStateButton)
     p_OnStateButton->AbortPlayback();
-  GOOrganModel::GOSoundOrganInterfaceProxy::Disconnect();
+  GOOrganModel::GOSoundSamplerPlayerProxy::Disconnect();
   GOOrganModel::SetMidi(nullptr, nullptr);
   m_midi = NULL;
