+++ src/grandorgue/sound/GOSoundOrganEngine.h
@@ -41,8 +41,8 @@ class GOWindchest;
  *
  *   1. Constructor: GOSoundOrganEngine(organModel, memoryPool)
  *   2. Configuration: SetFromConfig(config) or manual setters
- *   3. BuildEngine(audioOutputConfigs, nSamplesPerBuffer, sampleRate,
- *      recorder) — builds tasks and the sampler pool. IDLE → BUILT.
+ *   3. BuildEngine(audioOutputConfigs, nSamplesPerBuffer, sampleRate) —
+ *      builds tasks and the sampler pool. IDLE → BUILT.
  *   4. StartEngine() — starts dispatching work to worker threads. BUILT →
  *      WORKING.
  *      ... ProcessAudioCallback() is called from the audio thread ...
@@ -255,7 +255,7 @@ public:
    */
 
   /** Returns a reference to the sampler player (used to connect the organ
-   * model via GOSoundOrganInterfaceProxy). */
+   * model via GOSoundSamplerPlayerProxy). */
   GOSoundSamplerPlayer &GetSamplerPlayer() { return m_SamplerPlayer; }
 
   /** @return GOSoundSamplerPlayer::UsedSamplerCount() — see there. */
