+++ src/grandorgue/sound/GOSoundSystem.h
@@ -13,10 +13,10 @@
 
 #include <wx/string.h>
 
+#include "interfaces/GOSoundCallbackConnector.h"
+#include "interfaces/GOSoundCloseListener.h"
 #include "threading/GOMutex.h"
 
-#include "GOSoundCallbackConnector.h"
-#include "GOSoundCloseListener.h"
 #include "GOSoundDevInfo.h"
 
 class GOConfig;
