+++ src/tests/testing/sound/GOTestSoundOrganEngineBase.h
@@ -10,7 +10,7 @@
 
 #include "GOTest.h"
 
-#include "sound/GOSoundCallbackConnector.h"
+#include "sound/interfaces/GOSoundCallbackConnector.h"
 
 class GOSoundOrganEngine;
